pyproject.toml
[project]
name = "saler"
version = "0.1.0"
description = "Ai Saler"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-protocol>=0.1.13",
    "aiofile>=3.9.0",
    "aiosqlite>=0.22.1",
    "dashscope>=1.25.13",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.66.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "sqlite-vss>=0.1.2",
    "sqlmodel>=0.0.37",
    "tenacity>=9.1.4",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "src"
module-root = ""

[project.scripts]
app = "src.main:run"
